Fixed an issue with key/value coding in expressions. Basically, one bug fix, one enhancement.
The bug fix was that we weren't doing a check to see of the "value" was an AnyObject, and then sending it a value(forKey:), which means we weren't resolving quite a few things. I should add a better unit test for that.

The enhancement is that when a literal is encountered, and it's a key path, we try to resolve the root of the key path against the execution context, and if it resolves, then we continue from there. This is nice, because we can then access key paths off of variables defined in the store, which, frankly, is kind of expected behavior.

I also changed how colors are gotten, because I didn't like that I was defining a whole bunch of top leve symbols that were basically "un-scoped". To get around that, I added an "object" variable type that instantiates an instance of an object. This, of course, can then be registered as a constant in the root store, so I defined that AJRVariableTypeColor be instantiated this way with the name "colors". Then, with the aforementioned enhancement to key/value coding, we can say "colors.black", and we'll resolve the key path first to the "colors" object in the store, and then to it's key, "black".