Enhanced Inspectors
There's two changes.

First, I fixed the issue with the swapping content on object inspectors. When the content is dynamic with selected object, it now won't be bound earlier (causing issues), but when it's needed. Basically this is done by deferring the creation of the slice until it's needed.

Secondly, labels can now contain markdown. This may be expensive, so I may do some optimizations for it, but for now, it seems to be reasonably performant. Latter profiling should determine if that's true or not.